fix: wip, adopting core implementation
