+++ backend/core/types/src/backend-swagger.ts
@@ -1259,6 +1259,8 @@ export class ListingsService {
       orderBy?: OrderByFieldsEnum
       /**  */
       orderDir?: OrderDirEnum
+      /**  */
+      search?: string
     } = {} as any,
     options: IRequestOptions = {}
   ): Promise<PaginatedListing> {
@@ -1273,6 +1275,7 @@ export class ListingsService {
         view: params["view"],
         orderBy: params["orderBy"],
         orderDir: params["orderDir"],
+        search: params["search"],
       }
       let data = null
