+++ sites/partners/pages/index.tsx
@@ -1,237 +1,171 @@
-import React, { useMemo, useContext, useEffect, useRef, useState, useCallback } from "react"
+import React, { useMemo, useContext } from "react"
 import Head from "next/head"
 import {
-  PageHeader,
+  //   NavigationHeader,
   t,
-  AuthContext,
   Button,
   LocalizedLink,
-  AgPagination,
-  AG_PER_PAGE_OPTIONS,
-  LoadingOverlay,
+  AgTable,
+  AuthContext,
+  useAgTable,
+  PageHeader,
 } from "@bloom-housing/ui-components"
-import { AgGridReact } from "ag-grid-react"
-import { ColumnApi, ColumnState, GridOptions } from "ag-grid-community"
-import { useRouter } from "next/router"
+import dayjs from "dayjs"
+import { ColDef, ColGroupDef } from "ag-grid-community"
 import { useListingsData } from "../lib/hooks"
 import Layout from "../layouts"
 import { MetaTags } from "../src/MetaTags"
-import { ListingStatus, OrderByFieldsEnum, OrderDirEnum } from "@bloom-housing/backend-core/types"
+class formatLinkCell {
+  link: HTMLAnchorElement
 
-const COLUMN_STATE_KEY = "column-state"
+  init(params) {
+    this.link = document.createElement("a")
+    this.link.classList.add("text-blue-700")
+    this.link.innerText = params.valueFormatted || params.value
+    this.link.setAttribute("href", `/listings/${params.data.id}/applications`)
+  }
 
-type ListingListSortOptions = {
-  orderBy: OrderByFieldsEnum
-  orderDir: OrderDirEnum
+  getGui() {
+    return this.link
+  }
+}
+
+class formatWaitlistStatus {
+  text: HTMLSpanElement
+
+  init({ data }) {
+    const isWaitlistOpen = data.waitlistCurrentSize < data.waitlistMaxSize
+
+    this.text = document.createElement("span")
+    this.text.innerHTML = isWaitlistOpen ? t("t.yes") : t("t.no")
+  }
+
+  getGui() {
+    return this.text
+  }
+}
+
+class ApplicationsLink extends formatLinkCell {
+  init(params) {
+    super.init(params)
+    this.link.setAttribute("href", `/listings/${params.data.id}/applications`)
+    this.link.setAttribute("data-test-id", "listing-status-cell")
+  }
+}
+
+class ListingsLink extends formatLinkCell {
+  init(params) {
+    super.init(params)
+    this.link.setAttribute("href", `/listings/${params.data.id}`)
+  }
 }
 
 export default function ListingsList() {
+  const metaDescription = t("pageDescription.welcome", { regionName: t("region.name") })
+
   const { profile } = useContext(AuthContext)
   const isAdmin = profile.roles?.isAdmin || false
 
-  const router = useRouter()
+  const tableOptions = useAgTable()
 
-  const [gridColumnApi, setGridColumnApi] = useState<ColumnApi | null>(null)
-
-  /* Pagination */
-  const [itemsPerPage, setItemsPerPage] = useState<number>(AG_PER_PAGE_OPTIONS[0])
-  const [currentPage, setCurrentPage] = useState<number>(1)
-
-  /* OrderBy columns */
-  const [sortOptions, setSortOptions] = useState<ListingListSortOptions>({
-    orderBy: OrderByFieldsEnum.name,
-    orderDir: OrderDirEnum.ASC,
-  })
-
-  const metaDescription = t("pageDescription.welcome", { regionName: t("region.name") })
-  const metaImage = "" // TODO: replace with hero image
-
-  function saveColumnState(api: ColumnApi) {
-    const columnState = api.getColumnState()
-    const columnStateJSON = JSON.stringify(columnState)
-    sessionStorage.setItem(COLUMN_STATE_KEY, columnStateJSON)
+  const gridComponents = {
+    ApplicationsLink,
+    formatLinkCell,
+    formatWaitlistStatus,
+    ListingsLink,
   }
-
-  function onGridReady(params) {
-    setGridColumnApi(params.columnApi)
-  }
-
-  class formatLinkCell {
-    link: HTMLAnchorElement
-
-    init(params) {
-      this.link = document.createElement("a")
-      this.link.classList.add("text-blue-700")
-      this.link.innerText = params.valueFormatted || params.value
-      this.link.addEventListener("click", function () {
-        void saveColumnState(params.columnApi)
-        void router.push(`/listings/${params.data.id}`)
-      })
-    }
-
-    getGui() {
-      return this.link
-    }
-  }
-
-  class formatWaitlistStatus {
-    text: HTMLSpanElement
-
-    init({ data }) {
-      const isWaitlistOpen = data.waitlistCurrentSize < data.waitlistMaxSize
-
-      this.text = document.createElement("span")
-      this.text.innerHTML = isWaitlistOpen ? t("t.yes") : t("t.no")
-    }
-
-    getGui() {
-      return this.text
-    }
-  }
-
-  /* Grid Functionality and Formatting */
-  // update table items order on sort change
-  const initialLoadOnSort = useRef<boolean>(false)
-  const onSortChange = useCallback((columns: ColumnState[]) => {
-    // prevent multiple fetch on initial render
-    if (!initialLoadOnSort.current) {
-      initialLoadOnSort.current = true
-      return
-    }
-
-    const sortedBy = columns.find((col) => col.sort)
-    const { colId, sort } = sortedBy || {}
-
-    let col = colId
-    if (colId === "isVerified") {
-      col = "verified"
-    }
-    const allowedSortColIds: string[] = Object.values(OrderByFieldsEnum)
-    if (allowedSortColIds.includes(col)) {
-      const name = OrderByFieldsEnum[col]
-      setSortOptions({
-        orderBy: name,
-        orderDir: sort.toUpperCase() as OrderDirEnum,
-      })
-    }
-  }, [])
 
   const columnDefs = useMemo(() => {
-    const columns = [
+    const columns: (ColDef | ColGroupDef)[] = [
       {
         headerName: t("listings.listingName"),
         field: "name",
         sortable: true,
-        unSortIcon: true,
-        sort: "asc",
         filter: false,
-        width: 450,
-        minWidth: 100,
-        cellRenderer: "formatLinkCell",
-      },
-      {
-        headerName: t("listings.buildingAddress"),
-        field: "buildingAddress.street",
-        sortable: false,
-        filter: false,
-        width: 495,
-        minWidth: 100,
-        valueFormatter: ({ value }) => (value ? value : t("t.none")),
+        resizable: true,
+        cellRenderer: "ListingsLink",
       },
       {
         headerName: t("listings.listingStatusText"),
         field: "status",
-        sortable: true,
-        unSortIcon: true,
+        sortable: false,
         filter: false,
-        width: 150,
-        minWidth: 100,
-        valueFormatter: ({ value }) => {
-          switch (value) {
-            case ListingStatus.active:
-              return t("t.public")
-            case ListingStatus.pending:
-              return t("t.draft")
-            case ListingStatus.closed:
-              return t("listings.closed")
-            default:
-              return ""
-          }
-        },
+        resizable: true,
+        flex: 1,
+        valueFormatter: ({ value }) => t(`listings.listingStatus.${value}`),
+        cellRenderer: "ApplicationsLink",
       },
       {
-        headerName: t("listings.verified"),
-        field: "isVerified",
-        sortable: true,
-        unSortIcon: true,
+        headerName: t("listings.applicationDeadline"),
+        field: "applicationDueDate",
+        sortable: false,
         filter: false,
-        width: 150,
-        minWidth: 100,
-        valueFormatter: ({ value }) => (value ? t("t.yes") : t("t.no")),
+        resizable: true,
+        valueFormatter: ({ value }) => (value ? dayjs(value).format("MM/DD/YYYY") : t("t.none")),
+      },
+      {
+        headerName: t("listings.availableUnits"),
+        field: "unitsAvailable",
+        sortable: false,
+        filter: false,
+        resizable: true,
+      },
+      {
+        headerName: t("listings.waitlist.open"),
+        field: "waitlistCurrentSize",
+        sortable: false,
+        filter: false,
+        resizable: true,
+        cellRenderer: "formatWaitlistStatus",
       },
     ]
     return columns
   }, [])
 
   const { listingDtos, listingsLoading } = useListingsData({
-    page: currentPage,
-    limit: itemsPerPage,
-    listingIds: !isAdmin
-      ? profile?.leasingAgentInListings?.map((listing) => listing.id)
-      : undefined,
-    orderBy: sortOptions.orderBy,
-    orderDir: sortOptions.orderDir,
-    view: "partnerList",
+    page: tableOptions.pagination.currentPage,
+    limit: tableOptions.pagination.itemsPerPage,
+    search: tableOptions.filter.filterValue,
+    // search: tableOptions.search
+    // userId: !isAdmin ? profile?.id : undefined,
+    // sort: tableOptions.sort.sortOptions,
   })
-
-  const gridOptions: GridOptions = {
-    onSortChanged: (params) => {
-      saveColumnState(params.columnApi)
-      onSortChange(params.columnApi.getColumnState())
-    },
-    components: {
-      formatLinkCell,
-      formatWaitlistStatus,
-    },
-    suppressNoRowsOverlay: listingsLoading,
-  }
-
-  /* Pagination */
-  // reset page to 1 when user change limit
-  useEffect(() => {
-    setCurrentPage(1)
-  }, [itemsPerPage])
-
-  /* Data Performance */
-  // Load a table state on initial render & pagination change (because the new data comes from the API)
-  useEffect(() => {
-    const savedColumnState = sessionStorage.getItem(COLUMN_STATE_KEY)
-    if (gridColumnApi && savedColumnState) {
-      const parsedState: ColumnState[] = JSON.parse(savedColumnState)
-
-      gridColumnApi.applyColumnState({
-        state: parsedState,
-        applyOrder: true,
-      })
-    }
-  }, [gridColumnApi, currentPage])
-
+  console.log(listingDtos)
   return (
     <Layout>
       <Head>
         <title>{t("nav.siteTitlePartners")}</title>
       </Head>
-      <MetaTags
-        title={t("nav.siteTitlePartners")}
-        image={metaImage}
-        description={metaDescription}
-      />
+      <MetaTags title={t("nav.siteTitlePartners")} description={metaDescription} />
       <PageHeader title={t("nav.listings")} className={"md:pt-16"} />
       <section>
         <article className="flex-row flex-wrap relative max-w-screen-xl mx-auto py-8 px-4">
-          <div className="ag-theme-alpine ag-theme-bloom">
-            <div className="flex justify-between">
-              <div className="w-56"></div>
+          <AgTable
+            id="listings-table"
+            pagination={{
+              perPage: tableOptions.pagination.itemsPerPage,
+              setPerPage: tableOptions.pagination.setItemsPerPage,
+              currentPage: tableOptions.pagination.currentPage,
+              setCurrentPage: tableOptions.pagination.setCurrentPage,
+            }}
+            config={{
+              gridComponents,
+              columns: columnDefs,
+              totalItemsLabel: t("listings.totalListings"),
+            }}
+            data={{
+              items: listingDtos?.items,
+              loading: listingsLoading,
+              totalItems: listingDtos?.meta.totalItems,
+              totalPages: listingDtos?.meta.totalPages,
+            }}
+            search={{
+              setSearch: tableOptions.filter.setFilterValue,
+            }}
+            sort={{
+              setSort: tableOptions.sort.setSortOptions,
+            }}
+            headerContent={
               <div className="flex-row">
                 {isAdmin && (
                   <LocalizedLink href={`/listings/add`}>
@@ -241,40 +175,8 @@ export default function ListingsList() {
                   </LocalizedLink>
                 )}
               </div>
-            </div>
-
-            <div className="applications-table mt-5">
-              <LoadingOverlay isLoading={listingsLoading}>
-                <AgGridReact
-                  onGridReady={onGridReady}
-                  gridOptions={gridOptions}
-                  defaultColDef={{
-                    resizable: true,
-                  }}
-                  columnDefs={columnDefs}
-                  rowData={listingDtos?.items}
-                  domLayout={"autoHeight"}
-                  headerHeight={83}
-                  rowHeight={58}
-                  suppressPaginationPanel={true}
-                  paginationPageSize={AG_PER_PAGE_OPTIONS[0]}
-                  suppressScrollOnNewData={true}
-                ></AgGridReact>
-              </LoadingOverlay>
-
-              <AgPagination
-                totalItems={listingDtos?.meta?.totalItems}
-                totalPages={listingDtos?.meta?.totalPages}
-                currentPage={currentPage}
-                itemsPerPage={itemsPerPage}
-                quantityLabel={t("listings.totalListings")}
-                setCurrentPage={setCurrentPage}
-                setItemsPerPage={setItemsPerPage}
-                onPerPageChange={() => setCurrentPage(1)}
-                includeBorder={true}
-              />
-            </div>
-          </div>
+            }
+          />
         </article>
       </section>
     </Layout>
