--- sites/partners/pages/index2.tsx
@@ -1,194 +1,0 @@
-import React, { useMemo, useContext, useState } from "react"
-import Head from "next/head"
-import {
-  //   NavigationHeader,
-  t,
-  Button,
-  LocalizedLink,
-  AgTable,
-  useAgTable,
-  AuthContext,
-} from "@bloom-housing/ui-components"
-import dayjs from "dayjs"
-import { ColDef, ColGroupDef, ColumnApi } from "ag-grid-community"
-import { useListingsData } from "../lib/hooks"
-import Layout from "../layouts"
-import { MetaTags } from "../src/MetaTags"
-import router from "next/router"
-
-export default function ListingsList() {
-  const metaDescription = t("pageDescription.welcome", { regionName: t("region.name") })
-
-  const { profile } = useContext(AuthContext)
-  const isAdmin = profile.roles?.isAdmin || false
-  const [gridColumnApi, setGridColumnApi] = useState<ColumnApi | null>(null)
-
-  const tableOptions = useAgTable()
-
-  function saveColumnState(api: ColumnApi) {
-    const columnState = api.getColumnState()
-    const columnStateJSON = JSON.stringify(columnState)
-    sessionStorage.setItem(COLUMN_STATE_KEY, columnStateJSON)
-  }
-
-  class formatLinkCell {
-    link: HTMLAnchorElement
-
-    init(params) {
-      this.link = document.createElement("a")
-      this.link.classList.add("text-blue-700")
-      this.link.innerText = params.valueFormatted || params.value
-      this.link.addEventListener("click", function () {
-        void saveColumnState(params.columnApi)
-        void router.push(`/listings/${params.data.id}`)
-      })
-    }
-
-    getGui() {
-      return this.link
-    }
-  }
-
-  class formatWaitlistStatus {
-    text: HTMLSpanElement
-
-    init({ data }) {
-      const isWaitlistOpen = data.waitlistCurrentSize < data.waitlistMaxSize
-
-      this.text = document.createElement("span")
-      this.text.innerHTML = isWaitlistOpen ? t("t.yes") : t("t.no")
-    }
-
-    getGui() {
-      return this.text
-    }
-  }
-
-  class ApplicationsLink extends formatLinkCell {
-    init(params) {
-      super.init(params)
-      this.link.setAttribute("href", `/listings/${params.data.id}/applications`)
-      this.link.setAttribute("data-test-id", "listing-status-cell")
-    }
-  }
-
-  class ListingsLink extends formatLinkCell {
-    init(params) {
-      super.init(params)
-      this.link.setAttribute("href", `/listings/${params.data.id}`)
-    }
-  }
-
-  const gridComponents = {
-    ApplicationsLink,
-    formatLinkCell,
-    formatWaitlistStatus,
-    ListingsLink,
-  }
-
-  const columnDefs = useMemo(() => {
-    const columns: (ColDef | ColGroupDef)[] = [
-      {
-        headerName: t("listings.listingName"),
-        field: "name",
-        sortable: true,
-        filter: false,
-        resizable: true,
-        cellRenderer: "ListingsLink",
-      },
-      {
-        headerName: t("listings.listingStatusText"),
-        field: "status",
-        sortable: false,
-        filter: false,
-        resizable: true,
-        flex: 1,
-        valueFormatter: ({ value }) => t(`listings.listingStatus.${value}`),
-        cellRenderer: "ApplicationsLink",
-      },
-      {
-        headerName: t("listings.applicationDeadline"),
-        field: "applicationDueDate",
-        sortable: false,
-        filter: false,
-        resizable: true,
-        valueFormatter: ({ value }) => (value ? dayjs(value).format("MM/DD/YYYY") : t("t.none")),
-      },
-      {
-        headerName: t("listings.availableUnits"),
-        field: "unitsAvailable",
-        sortable: false,
-        filter: false,
-        resizable: true,
-      },
-      {
-        headerName: t("listings.waitlist.open"),
-        field: "waitlistCurrentSize",
-        sortable: false,
-        filter: false,
-        resizable: true,
-        cellRenderer: "formatWaitlistStatus",
-      },
-    ]
-    return columns
-  }, [])
-
-  const { listingDtos, listingsLoading } = useListingsData({
-    page: tableOptions.pagination.currentPage,
-    limit: tableOptions.pagination.itemsPerPage,
-    search: tableOptions.filter.filterValue,
-    userId: !isAdmin ? profile?.id : undefined,
-    sort: tableOptions.sort.sortOptions,
-  })
-
-  return (
-    <Layout>
-      <Head>
-        <title>{t("nav.siteTitlePartners")}</title>
-      </Head>
-      <MetaTags title={t("nav.siteTitlePartners")} description={metaDescription} />
-      <NavigationHeader title={t("nav.listings")} />
-      <section>
-        <article className="flex-row flex-wrap relative max-w-screen-xl mx-auto py-8 px-4">
-          <AgTable
-            id="listings-table"
-            pagination={{
-              perPage: tableOptions.pagination.itemsPerPage,
-              setPerPage: tableOptions.pagination.setItemsPerPage,
-              currentPage: tableOptions.pagination.currentPage,
-              setCurrentPage: tableOptions.pagination.setCurrentPage,
-            }}
-            config={{
-              gridComponents,
-              columns: columnDefs,
-              totalItemsLabel: t("listings.totalListings"),
-            }}
-            data={{
-              items: listingDtos?.items,
-              loading: listingsLoading,
-              totalItems: listingDtos?.meta.totalItems,
-              totalPages: listingDtos?.meta.totalPages,
-            }}
-            search={{
-              setSearch: tableOptions.filter.setFilterValue,
-            }}
-            sort={{
-              setSort: tableOptions.sort.setSortOptions,
-            }}
-            headerContent={
-              <div className="flex-row">
-                {isAdmin && (
-                  <LocalizedLink href={`/listings/add`}>
-                    <Button className="mx-1" onClick={() => false}>
-                      {t("listings.addListing")}
-                    </Button>
-                  </LocalizedLink>
-                )}
-              </div>
-            }
-          />
-        </article>
-      </section>
-    </Layout>
-  )
-}
